src/requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.6
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0
django-simple-history==3.1.1
drf-nested-routers==0.93.4
django-cors-headers==3.13.0
requests==2.28.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
django-debug-toolbar==3.5.0
parameterized~=0.8.1